refactor(settings): extract FeaturesSection to sections/
Moves the Features tab — editor-feature toggles (slash commands, wiki
links, tags), navigation (quick switcher, backlinks), and right-panel
widgets (calendar, timeline) — into its own file.

No behavior changes.
